+++ internal/storage/metricstore/prometheus/options.go
@@ -12,7 +12,7 @@ import (
 	"github.com/spf13/viper"
 
 	"github.com/jaegertracing/jaeger/internal/config/tlscfg"
-	"github.com/jaegertracing/jaeger/pkg/prometheus/config"
+	"github.com/jaegertracing/jaeger/internal/prometheus/config"
 )
 
 const (
